# proto/public_project/app.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/public_project/app.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AppService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/public_project/rpc/v1/hello": {
      "post": {
        "summary": "测试登入接口",
        "operationId": "AppService_Hello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/public_projectHelloRsp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/public_projectHelloReq"
            }
          }
        ],
        "tags": [
          "out"
        ]
      }
    },
    "/public_project/v1/get_user_info": {
      "post": {
        "summary": "获取用户信息",
        "operationId": "AppService_GetUserInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/public_projectGetUserInfoRep"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/public_projectGetUserInfoReq"
            }
          }
        ],
        "tags": [
          "out"
        ]
      }
    },
    "/public_project/v1/login": {
      "post": {
        "summary": "登录接口",
        "operationId": "AppService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/public_projectLoginRep"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/public_projectLoginReq"
            }
          }
        ],
        "tags": [
          "out"
        ]
      }
    },
    "/public_project/v1/logout": {
      "post": {
        "summary": "登出接口",
        "operationId": "AppService_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/public_projectLogoutRep"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/public_projectLogoutReq"
            }
          }
        ],
        "tags": [
          "out"
        ]
      }
    },
    "/public_project/v1/refresh_token": {
      "post": {
        "summary": "登录令牌",
        "operationId": "AppService_RefreshToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/public_projectRefreshTokenRep"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/public_projectRefreshTokenReq"
            }
          }
        ],
        "tags": [
          "out"
        ]
      }
    }
  },
  "definitions": {
    "public_projectGetUserInfoRep": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/public_projectUser"
        }
      }
    },
    "public_projectGetUserInfoReq": {
      "type": "object"
    },
    "public_projectHelloReq": {
      "type": "object"
    },
    "public_projectHelloRsp": {
      "type": "object"
    },
    "public_projectLoginRep": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "refresh_token": {
          "type": "string"
        },
        "expires_in": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "public_projectLoginReq": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "public_projectLogoutRep": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "public_projectLogoutReq": {
      "type": "object",
      "properties": {
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "public_projectRefreshTokenRep": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "expires_in": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "public_projectRefreshTokenReq": {
      "type": "object",
      "properties": {
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "public_projectUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "user_name": {
          "type": "string"
        },
        "user_account": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "created_time": {
          "type": "string",
          "format": "int64"
        },
        "updated_time": {
          "type": "string",
          "format": "int64"
        },
        "is_deleted": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "用户信息"
    }
  }
}